[#38] refactor: 영수증과 리뷰 정보를 보여주는 메소드를 ManagerDAO로 이동
