+++ BE/src/main/java/com/codesquad/airbnb/room/infra/RoomDAO.java
@@ -126,37 +126,4 @@ public class RoomDAO {
         OptionalDouble average = prices.stream().mapToDouble(a -> a).average();
         return average.isPresent() ? (int) average.getAsDouble() : 0;
     }
-
-    public Confirmation showBillAndReview(ManagerDAO managerDAO, Long roomId, ReservationDate reservationDate, Guest guest) {
-
-        if(!managerDAO.canReserve(roomId, reservationDate.getCheckInDate(), reservationDate.getCheckOutDate())) {
-            throw new IllegalArgumentException("Already reserved room, Please reserve another room!");
-        }
-
-        String sql = "SELECT r.room_id AS id, IF(r.host_is_superhost = 't', p.price * 0.9, p.price) AS price, p.cleaning_fee, p.security_deposit, r2.number_of_reviews, r2.review_scores_rating AS rating FROM rooms r INNER JOIN prices p on r.room_id = p.room_id INNER JOIN reviews r2 on r.room_id = r2.room_id WHERE r.room_id = ? GROUP BY r.room_id;";
-
-        RowMapper<Confirmation> reservationRowMapper = new RowMapper<Confirmation>() {
-            @Override
-            public Confirmation mapRow(ResultSet rs, int rowNum) throws SQLException {
-                int salesPrice = rs.getInt("price");
-                int length = (int) (ChronoUnit.DAYS.between(reservationDate.getCheckInDate(), reservationDate.getCheckOutDate()));
-                int cleaningFee = rs.getInt("cleaning_fee");
-                int deposit = rs.getInt("security_deposit");
-                int priceWithFee = (salesPrice * length) + cleaningFee + deposit;
-
-                Bill bill = new Bill(
-                        salesPrice,
-                        length + 1,
-                        salesPrice * length,
-                        cleaningFee,
-                        deposit,
-                        priceWithFee
-                );
-                Review review = new Review(rs.getInt("number_of_reviews"), rs.getDouble("rating"));
-                return new Confirmation(bill, review);
-            }
-        };
-
-        return this.jdbcTemplate.queryForObject(sql, new Object[]{roomId}, reservationRowMapper);
-    }
 }
